Move definitions of API objects to api_objects.hpp
