+++ libraries/app/include/graphene/app/api_objects.hpp
@@ -24,15 +24,21 @@
 #pragma once
 
 #include <graphene/chain/account_object.hpp>
+#include <graphene/chain/asset_object.hpp>
 #include <graphene/chain/vesting_balance_object.hpp>
-#include <graphene/chain/market_evaluator.hpp>
 #include <graphene/chain/market_object.hpp>
 #include <graphene/chain/proposal_object.hpp>
 #include <graphene/chain/withdraw_permission_object.hpp>
 #include <graphene/chain/htlc_object.hpp>
 
+#include <graphene/api_helper_indexes/api_helper_indexes.hpp>
+#include <graphene/market_history/market_history_plugin.hpp>
+
+#include <fc/optional.hpp>
+
 namespace graphene { namespace app {
    using namespace graphene::chain;
+   using namespace graphene::market_history;
 
    struct more_data
    {
@@ -72,6 +78,74 @@ namespace graphene { namespace app {
       more_data                        more_data_available;
    };
 
+   struct order
+   {
+      string                     price;
+      string                     quote;
+      string                     base;
+   };
+
+   struct order_book
+   {
+     string                      base;
+     string                      quote;
+     vector< order >             bids;
+     vector< order >             asks;
+   };
+
+   struct market_ticker
+   {
+      time_point_sec             time;
+      string                     base;
+      string                     quote;
+      string                     latest;
+      string                     lowest_ask;
+      string                     highest_bid;
+      string                     percent_change;
+      string                     base_volume;
+      string                     quote_volume;
+
+      market_ticker() {}
+      market_ticker(const market_ticker_object& mto,
+                    const fc::time_point_sec& now,
+                    const asset_object& asset_base,
+                    const asset_object& asset_quote,
+                    const order_book& orders);
+      market_ticker(const fc::time_point_sec& now,
+                    const asset_object& asset_base,
+                    const asset_object& asset_quote);
+   };
+
+   struct market_volume
+   {
+      time_point_sec             time;
+      string                     base;
+      string                     quote;
+      string                     base_volume;
+      string                     quote_volume;
+   };
+
+   struct market_trade
+   {
+      int64_t                    sequence = 0;
+      fc::time_point_sec         date;
+      string                     price;
+      string                     amount;
+      string                     value;
+      account_id_type            side1_account_id = GRAPHENE_NULL_ACCOUNT;
+      account_id_type            side2_account_id = GRAPHENE_NULL_ACCOUNT;
+   };
+
+   struct extended_asset_object : asset_object
+   {
+      extended_asset_object() {}
+      explicit extended_asset_object( const asset_object& a ) : asset_object( a ) {}
+      explicit extended_asset_object( asset_object&& a ) : asset_object( std::move(a) ) {}
+
+      optional<share_type> total_in_collateral;
+      optional<share_type> total_backing_collateral;
+   };
+
 } }
 
 FC_REFLECT( graphene::app::more_data,
@@ -100,3 +174,13 @@ FC_REFLECT( graphene::app::full_account,
             (htlcs_to)
             (more_data_available)
           )
+
+FC_REFLECT( graphene::app::order, (price)(quote)(base) );
+FC_REFLECT( graphene::app::order_book, (base)(quote)(bids)(asks) );
+FC_REFLECT( graphene::app::market_ticker,
+            (time)(base)(quote)(latest)(lowest_ask)(highest_bid)(percent_change)(base_volume)(quote_volume) );
+FC_REFLECT( graphene::app::market_volume, (time)(base)(quote)(base_volume)(quote_volume) );
+FC_REFLECT( graphene::app::market_trade, (sequence)(date)(price)(amount)(value)(side1_account_id)(side2_account_id) );
+
+FC_REFLECT_DERIVED( graphene::app::extended_asset_object, (graphene::chain::asset_object),
+                    (total_in_collateral)(total_backing_collateral) );
